Add missing imports of activity.cpp
