+++ src/command.cpp
@@ -1,5 +1,6 @@
 #include "command.hpp"
 #include "ability.hpp"
+#include "activity.hpp"
 #include "animation.hpp"
 #include "audio.hpp"
 #include "buff.hpp"
