+++ src/food.hpp
@@ -6,9 +6,6 @@ namespace elona
 
 enum class curse_state_t;
 
-void continuous_action_eating();
-void continuous_action_eating_finish();
-
 int chara_anorexia(int = 0);
 void chara_vomit(int = 0);
 void eatstatus(curse_state_t, int);
@@ -22,7 +19,7 @@ void cook();
 
 void make_dish(int, int);
 
-void apply_general_eating_effect();
+void apply_general_eating_effect(int);
 void eating_effect_eat_iron();
 void eating_effect_insanity();
 void eating_effect_eat_horse();
