+++ src/magic.cpp
@@ -1,4 +1,5 @@
 #include "ability.hpp"
+#include "activity.hpp"
 #include "animation.hpp"
 #include "audio.hpp"
 #include "buff.hpp"
